requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --no-emit-index-url requirements.in
#
asgiref==3.7.2
    # via django
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
django==3.2.23
    # via
    #   -r requirements.in
    #   djangorestframework
    #   drf-spectacular
djangorestframework==3.12.4
    # via
    #   -r requirements.in
    #   drf-spectacular
drf-spectacular==0.15.1
    # via -r requirements.in
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.21.1
    # via drf-spectacular
jsonschema-specifications==2023.12.1
    # via jsonschema
pillow==8.2.0
    # via -r requirements.in
psycopg2-binary==2.9.9
    # via -r requirements.in
pytz==2023.3.post1
    # via django
pyyaml==6.0.1
    # via drf-spectacular
referencing==0.33.0
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.17.1
    # via
    #   jsonschema
    #   referencing
sqlparse==0.4.4
    # via django
uritemplate==4.1.1
    # via drf-spectacular
